# app/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/mealplans": {"post": {"tags": ["mealplans"], "summary": "Create Mealplan", "description": "Create a new meal plan.", "operationId": "create_mealplan_mealplans_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Mealplan"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Mealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["mealplans"], "summary": "Get All Mealplans", "description": "Retrieve all meal plans with pagination.", "operationId": "get_all_mealplans_mealplans_get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "description": "Number of records to skip", "default": 0, "title": "Skip"}, "description": "Number of records to skip"}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "description": "Number of records to retrieve", "default": 10, "title": "Limit"}, "description": "Number of records to retrieve"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/mealplans/{meal_id}": {"get": {"tags": ["mealplans"], "summary": "Get Mealplan By Id", "description": "Retrieve a meal plan by its ID.", "operationId": "get_mealplan_by_id_mealplans__meal_id__get", "parameters": [{"name": "meal_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Meal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Mealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["mealplans"], "summary": "Update Mealplan By Id", "description": "Update a meal plan by its ID.", "operationId": "update_mealplan_by_id_mealplans__meal_id__put", "parameters": [{"name": "meal_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Meal Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Mealplan"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Mealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["mealplans"], "summary": "Delete Mealplan By Id", "description": "Delete a meal plan by its ID.", "operationId": "delete_mealplan_by_id_mealplans__meal_id__delete", "parameters": [{"name": "meal_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Meal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/weekly-mealplans/{week_plan_id}": {"put": {"tags": ["weekly-mealplans"], "summary": "Update Weekly Mealplan By Id", "description": "Update a meal plan by its ID.", "operationId": "update_weekly_mealplan_by_id_weekly_mealplans__week_plan_id__put", "parameters": [{"name": "week_plan_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Week Plan Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WeeklyMealplan"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WeeklyMealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["weekly-mealplans"], "summary": "Get Weekly Mealplan By Id", "description": "Retrieve a weekly meal plan by its ID.", "operationId": "get_weekly_mealplan_by_id_weekly_mealplans__week_plan_id__get", "parameters": [{"name": "week_plan_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Week Plan Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WeeklyMealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/daily-mealplans/{day_plan_id}": {"put": {"tags": ["daily-mealplans"], "summary": "Update Daily Mealplan By Id", "description": "Update a meal plan by its ID.", "operationId": "update_daily_mealplan_by_id_daily_mealplans__day_plan_id__put", "parameters": [{"name": "day_plan_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Day Plan Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DailyMealplan"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DailyMealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["daily-mealplans"], "summary": "Delete Daily Mealplan By Id", "description": "Delete a daily meal plan by its ID.", "operationId": "delete_daily_mealplan_by_id_daily_mealplans__day_plan_id__delete", "parameters": [{"name": "day_plan_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Day Plan Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["daily-mealplans"], "summary": "Get Daily Mealplan By Id", "description": "Retrieve a daily meal plan by its ID.", "operationId": "get_daily_mealplan_by_id_daily_mealplans__day_plan_id__get", "parameters": [{"name": "day_plan_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Day Plan Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DailyMealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/weekly-mealplans/{weekly_meal_id}": {"delete": {"tags": ["weekly-mealplans"], "summary": "Delete Weekly Mealplan By Id", "description": "Delete a weekly meal plan by its ID.", "operationId": "delete_weekly_mealplan_by_id_weekly_mealplans__weekly_meal_id__delete", "parameters": [{"name": "weekly_meal_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Weekly Meal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/weekly-mealplans": {"post": {"tags": ["weekly-mealplans"], "summary": "Create Weekly Mealplan", "description": "Create a new weekly meal plan.", "operationId": "create_weekly_mealplan_weekly_mealplans_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WeeklyMealplan"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WeeklyMealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["weekly-mealplans"], "summary": "Get All Weekly Mealplans", "description": "Retrieve all meal plans with pagination.", "operationId": "get_all_weekly_mealplans_weekly_mealplans_get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "description": "Number of records to skip", "default": 0, "title": "Skip"}, "description": "Number of records to skip"}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "description": "Number of records to retrieve", "default": 10, "title": "Limit"}, "description": "Number of records to retrieve"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/daily-mealplans": {"post": {"tags": ["daily-mealplans"], "summary": "Create Daily Mealplan", "description": "Create a new daily meal plan.", "operationId": "create_daily_mealplan_daily_mealplans_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DailyMealplan"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DailyMealplan"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["daily-mealplans"], "summary": "Get All Daily Mealplans", "description": "Retrieve all meal plans with pagination.", "operationId": "get_all_daily_mealplans_daily_mealplans_get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "description": "Number of records to skip", "default": 0, "title": "Skip"}, "description": "Number of records to skip"}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "description": "Number of records to retrieve", "default": 10, "title": "Limit"}, "description": "Number of records to retrieve"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/weekly-mealplans/{week_plan_id}/daily-mealplans": {"get": {"tags": ["weekly-mealplans"], "summary": "Get Daily Mealplans By Week", "description": "Retrieve all daily meal plans within a weekly plan by the weekly plan ID.", "operationId": "get_daily_mealplans_by_week_weekly_mealplans__week_plan_id__daily_mealplans_get", "parameters": [{"name": "week_plan_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Week Plan Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/DailyMealplan"}, "title": "Response Get Daily Mealplans By Week Weekly Mealplans  Week Plan Id  Daily Mealplans Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/weekly-mealplans/": {"get": {"tags": ["weekly-mealplans"], "summary": "Get Daily Meal Plans By Date", "description": "Retrieve all daily meal plans within a weekly plan by the weekly plan ID.", "operationId": "get_daily_meal_plans_by_date_weekly_mealplans__get", "parameters": [{"name": "date", "in": "query", "required": true, "schema": {"type": "string", "title": "Date"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/WeeklyMealplan"}, "title": "Response Get Daily Meal Plans By Date Weekly Mealplans  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"DailyMealplan": {"properties": {"day_plan_id": {"type": "integer", "title": "Day Plan Id"}, "week_plan_id": {"type": "integer", "title": "Week Plan Id"}, "date": {"type": "string", "title": "Date"}, "meal_id": {"type": "integer", "title": "Meal Id"}, "links": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Links"}}, "type": "object", "required": ["day_plan_id", "week_plan_id", "date", "meal_id"], "title": "DailyMealplan", "example": {"date": "2024-10-01", "day_plan_id": 1, "links": {"delete": "/daily-mealplans/1/delete", "edit": "/daily-mealplans/1/edit", "mealplan": "/mealplans/10", "self": "/daily-mealplans/1", "week": "/weekly-mealplans/1"}, "meal_id": 10, "week_plan_id": 1}}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Mealplan": {"properties": {"meal_id": {"type": "integer", "title": "Meal Id"}, "breakfast_recipe": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Breakfast Recipe"}, "lunch_recipe": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Lunch Recipe"}, "dinner_recipe": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Dinner Recipe"}, "links": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Links"}}, "type": "object", "required": ["meal_id"], "title": "Mealplan", "example": {"breakfast_recipe": 171, "dinner_recipe": 192, "links": {"breakfast_recipe": "/recipes/171", "delete": "/mealplans/10/delete", "dinner_recipe": "/recipes/192", "edit": "/mealplans/10/edit", "lunch_recipe": "/recipes/180", "self": "/mealplans/10"}, "lunch_recipe": 180, "meal_id": 10}}, "PaginatedResponse": {"properties": {"items": {"items": {}, "type": "array", "title": "Items"}, "links": {"type": "object", "title": "Links"}}, "type": "object", "required": ["items", "links"], "title": "PaginatedResponse", "example": {"items": [{"date": "2024-10-01", "day_plan_id": 1, "meal_id": 10, "week_plan_id": 1}, {"date": "2024-10-02", "day_plan_id": 2, "meal_id": 11, "week_plan_id": 1}], "links": {"next": "/weekly-mealplans/1/daily-mealplans?page=2", "prev": "/weekly-mealplans/1/daily-mealplans?page=1", "self": "/weekly-mealplans/1/daily-mealplans?page=1"}}}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "WeeklyMealplan": {"properties": {"week_plan_id": {"type": "integer", "title": "Week Plan Id"}, "start_date": {"type": "string", "title": "Start Date"}, "end_date": {"type": "string", "title": "End Date"}, "links": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Links"}}, "type": "object", "required": ["week_plan_id", "start_date", "end_date"], "title": "WeeklyMealplan", "example": {"end_date": "2024-10-07", "links": {"daily_mealplans": "/weekly-mealplans/1/daily-mealplans", "delete": "/weekly-mealplans/1/delete", "edit": "/weekly-mealplans/1/edit", "self": "/weekly-mealplans/1"}, "start_date": "2024-10-01", "week_plan_id": 1}}}}}